feat(zimbra): add pro offer to the order page
ref: #PRDCOL-226
