+++ packages/manager/apps/zimbra/src/data/api/order/mocks.ts
@@ -74,6 +74,70 @@ export const orderCatalogMock: order.publicOrder.Catalog = {
         },
       },
     },
+    {
+      planCode: 'zimbra-account-pp-pro',
+      invoiceName: 'Zimbra account',
+      addonFamilies: [],
+      product: 'zimbra-account',
+      pricingType: order.cart.GenericProductPricingTypeEnum.RENTAL,
+      consumptionConfiguration: null,
+      pricings: [
+        {
+          phase: 0,
+          capacities: [
+            order.cart.GenericProductPricingCapacitiesEnum.INSTALLATION,
+            order.cart.GenericProductPricingCapacitiesEnum.RENEW,
+          ],
+          commitment: 0,
+          description: 'monthly pricing',
+          interval: 1,
+          intervalUnit: IntervalUnitType.month,
+          quantity: {
+            min: 1,
+            max: null,
+          },
+          repeat: {
+            min: 1,
+            max: null,
+          },
+          price: 0,
+          tax: 0,
+          mode: 'default',
+          strategy: order.cart.GenericProductPricingStrategyEnum.VOLUME,
+          mustBeCompleted: false,
+          type: order.cart.GenericProductPricingTypeEnum.RENTAL,
+          promotions: [],
+          engagementConfiguration: null,
+        },
+      ],
+      configurations: [],
+      family: null,
+      blobs: {
+        commercial: {
+          name: 'Zimbra Pro',
+          price: {
+            precision: 2,
+            interval: 'P1M',
+            unit: 'M/account',
+          },
+        },
+        meta: {
+          configurations: [
+            {
+              name: 'domain_choices',
+              values: [
+                {
+                  value: 'new_or_transfer',
+                },
+                {
+                  value: 'use_mine',
+                },
+              ],
+            },
+          ],
+        },
+      },
+    },
   ],
   products: [
     {
